# perf: don't reuse the same incompatibility repeatedly