This includes #87, the only new part is the last commit. It is intended to be merged after it.

Most of the time `relation` for an incompatibility (like `A` depends on `B`) is called at least twice, once when `A` gets added and when `B` gets added. After the first call we get `AlmostSatisfied` and change the `partial_solution` so that it will be contradicted the next time. This adds a HashSet to keep track of all the `IncompId`s we know will stay contradicted until the next backtrack. `IncompId` are just a small integer so hash very quickly.

It is possible to figure out witch `IncompId`s are still used after a backtrack, but so far nothing that is pulls its weight. 

```
Warning: Unable to complete 100 samples in 20.0s. You may wish to increase target time to 23.6s, or reduce sample count to 80.
large_cases/elm_str_SemanticVersion.ron
                        time:   [244.05 ms 244.40 ms 244.79 ms]
                        change: [-9.0431% -8.8780% -8.7172%] (p = 0.00 < 0.05)
                        Performance has improved.
large_cases/large_case_u16_NumberVersion.ron
                        time:   [9.3124 ms 9.3213 ms 9.3302 ms]
                        change: [-13.284% -13.164% -13.049%] (p = 0.00 < 0.05)
                        Performance has improved.
Benchmarking large_cases/zuse_str_SemanticVersion.ron: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 20.0s. You may wish to increase target time to 35.2s, or reduce sample count to 50.
large_cases/zuse_str_SemanticVersion.ron
                        time:   [351.46 ms 351.77 ms 352.08 ms]
                        change: [-12.531% -12.418% -12.310%] (p = 0.00 < 0.05)
                        Performance has improved.
```