@Eh2406 2021-05-06
There are more complicated structures that can support backtracking or filter out items faster than a `HashSet.contains`, but this was a simple change with a significant perf improvement. Love to play with a bigger PR, but only after the Range Trait.