@mpizenberg 2021-05-21
Indeed, I should have time this afternoon

On Fri, May 21, 2021, 04:24 Jacob Finkelman ***@***.***>
wrote:

> I am feeling the urge to try some of the more complicated options. Will
> you have time to take a look at this before I succumb?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pubgrub-rs/pubgrub/pull/88#issuecomment-845604890>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAWFOCMUCGV4UGQCVBY35JTTOW74LANCNFSM44DHLUFA>
> .
>