@Eh2406 2021-05-22
That is fabulous, thank you!
Do you think the rename should trigger a new 10 day open time or can we merge now?