@mpizenberg 2021-05-22
Nah I think nobody else is gonna complain if you merge this ^^

On Sat, May 22, 2021, 02:12 Jacob Finkelman ***@***.***>
wrote:

> That is fabulous, thank you!
> Do you think the rename should trigger a new 10 day open time or can we
> merge now?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pubgrub-rs/pubgrub/pull/88#issuecomment-846317382>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAWFOCOTZXJUXXLYX37SC33TO3ZITANCNFSM44DHLUFA>
> .
>